ai-lead-outreach/backend/requirements.txt
flask
flask-cors
mysql-connector-python
pandas
python-dotenv
google-generativeai
groq
openpyxl
ddgs
google-search-results
beautifulsoup4
playwright
requests
lxml
serpapi
